requirements.txt
Pillow==10.0.1
progress==1.5
pytest
pytest-cov
codecov
coverage
flask
numpy